Allowed the user to run it on themself
I had this in the Bloxlink Version, don’t know why I didn’t have it here. So I added it here, so basically if someone runs the command with no argument of the “argUser” the bot will make the argUser the msgObject.author (The user that ran the command).